\name{SpatialAcc-package}
\alias{SpatialAcc-package}
\alias{SpatialAcc}
\docType{package}
\title{Spatial Accessibility Measures}
\description{Provides a set of spatial accessibility measures from a set of locations (demand) to another set of locations (supply). It aims, among others, to support research on spatial accessibility to health care facilities. Includes the locations and some characteristics of major public hospitals in Greece.}

\details{
\tabular{ll}{
Package: \tab SpatialAcc\cr
Type: \tab Package\cr
Version: \tab 0.1-5\cr
Date: \tab 2024-04-02\cr
License: \tab GPL (>= 2)
}

}
\author{
Stamatis Kalogirou

Maintainer: Stamatis Kalogirou <stamatis.science@gmail.com>}

\references{
Kalogirou, S., and Foley, R. (2006). Health, place and Hanly: Modelling accessibility to hospitals in Ireland. Irish Geography, 39(1), 52-68.

Kalogirou, S. (2017). Spatial inequality in the accessibility to hospitals in Greece, The International Archives of the Photogrammetry, Remote Sensing and Spatial Information Sciences, XLII-4/W2, 91-94, https://doi.org/10.5194/isprs-archives-XLII-4-W2-91-2017.
}

\keyword{package}
\keyword{spatial analysis}
\keyword{accessibility}
\keyword{spatial inequalities}
